# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_signIn",
        "summary": "Log in with credentials",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_registerUser",
        "summary": "Register a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/user": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/movies": {
      "post": {
        "operationId": "MovieController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMovieDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "MovieController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/movies/{id}": {
      "get": {
        "operationId": "MovieController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "MovieController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMovieDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "MovieController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Movie-Assignment",
    "description": "CRUD Operation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "strongPassword123",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "strongPassword123",
            "description": "User password"
          },
          "username": {
            "type": "string",
            "example": "username_123",
            "description": "Username"
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ]
      },
      "CreateMovieDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Lagan",
            "description": "Name"
          },
          "rating": {
            "type": "number",
            "example": 4,
            "description": "Rating"
          },
          "cast": {
            "example": [
              {
                "name": "Abhishek",
                "role": "Hero"
              }
            ],
            "description": "Cast",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genre": {
            "type": "string",
            "example": "Danger",
            "description": "Genre"
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "example": "08/11/2023",
            "description": "Release Date"
          }
        },
        "required": [
          "name",
          "rating",
          "cast",
          "genre",
          "releaseDate"
        ]
      },
      "UpdateMovieDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Lagan",
            "description": "Name"
          },
          "rating": {
            "type": "number",
            "example": 4,
            "description": "Rating"
          },
          "cast": {
            "example": [
              {
                "name": "Abhishek",
                "role": "Hero"
              }
            ],
            "description": "Cast",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genre": {
            "type": "string",
            "example": "Danger",
            "description": "Genre"
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "example": "08/11/2023",
            "description": "Release Date"
          }
        },
        "required": [
          "name",
          "rating",
          "cast",
          "genre",
          "releaseDate"
        ]
      }
    }
  }
}